#5 split up the class tests into single files for better readability
